number api is working, need some cosmetic work

+++ app/src/main/java/com/levp/getdatabynet/numFactApi/NumFact.kt
@@ -2,7 +2,7 @@ package com.levp.getdatabynet.numFactApi
 
 data class NumFact(
     val text: String,
-    var number: Int,
+    var number: Long,
     var type: String
 )
